Fix local config
When running in a local dev env, e.g. running from source, the cluster
role was missing access to namespaces which is now required by the k8s
storage plugin.

Also we were still using the locally built porter image, instead of the
agent image we are now publishing with new porter builds.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>