Pin the default porter-agent to a specific version
Do not use "latest" or "canary" when defaulting the porter-agent. We
don't want to encourage people to use mutable tags as it can easily
cause breakages when we release new versions.

We will bump this value each time we update the operator to work with a
new version of Porter instead. This is just for the default, so users
can always override it with an AgentConfig.

Fixes #51

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>