Set fsGroup on porter agent pod
We had only tested our installation on KinD, which doesn't seem to have
the same file permissions on its default storage class. When we tried to
run the operator on Azure (AKS), we found that the agent couldn't write
to the mounted volume (/porter-shared).

The porter agent runs as nonroot, not root, as part of our hardening
process for IronBank. So when the volume is mounted, it needs to be
marked with the same user id for nonroot so that the container has
permission to write to the volume.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>