Add AgentAction controller
I've split the installation controller in two, adding a new controller
just to manage running the porter agent. The installation controller now
only is reponsible for creating an AgentAction resource. This triggers
the AgentAction controller to make a porter agent job to execute the
command from the AgentAction.

It's a bit of an extra layer of indirection but it will simplify
additional controllers since they don't need to manage the porter agent
job. It also enables users to request that the porter agent is run and
can specify arbitrary commands, such as invoke.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>