Add AgentAction controller (#73)
* Include manager image in bundle

Add an image reference to the manager image (which has the controllers)
to the bundle and then reference it when installing the operator.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Add AgentAction controller

I've split the installation controller in two, adding a new controller
just to manage running the porter agent. The installation controller now
only is reponsible for creating an AgentAction resource. This triggers
the AgentAction controller to make a porter agent job to execute the
command from the AgentAction.

It's a bit of an extra layer of indirection but it will simplify
additional controllers since they don't need to manage the porter agent
job. It also enables users to request that the porter agent is run and
can specify arbitrary commands, such as invoke.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Use a constant for the default porter agent version

Also update the default version to alpha.13 which is the most recent
release.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Sync go.mod

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Bump schemaversion used in test param/creds

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Review feedback

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>