Make test wait time configurable
The ginkgo tests often have to trigger the controller with a change,
like editing an installation resource, and then wait for the controller
to react and the resulting porter action jobs to finish.

Depending on if you already have the images pulled, and how fast your
machine is, this can take a pretty variable amount of time. I've made
the wait time configurable so that we can increase it on our CI build.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>