Scan the manager image with Trivy during CI
Scan the porter-operator-manager image with Trivy for pushes and pull requests and then report the results back to GitHub Code Scanning service.

Reports for the main branch are displayed in the Security tab of the repository under "Code Scanning Alerts". For pull requests, they are displayed inline on the files tab.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>