Use an arm64 compatible mongodb image on arm machines
The Bitnami Helm chart that we use to run mongo doesn't support arm which is mostly a problem for devs on new macs. I have used the bitnami-compat tool to create an arm version of their mongodb image that is compatible with the Bitnami chart.

When you run mage deploy on an arm machine, it will automatically patch the helm chart to use the compatible image. The image can be switched to a different one with the PORTER_MONGODB_IMAGE environment variable.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>