Only trigger the push event for the main branch
Dependabot PRs use branches that exist on the upstream repository. This is causing the push event for the build workflow to trigger but since Dependabot is not trusted, the push (build) workflow always fails due to lack of access. I've updated the push event to only trigger for the main branch to avoid the dependabot PRs always being flagged as failed.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>