Calculate the version correctly when releasing from tag
During the last release of Porter Core there was an issue where the
pipeline published the canary version instead of the finalized released.
This happens because the build scripts calculate the version incorrectly
when multiple tags are present on the same commit.

The release process of the operator is slightly different, but updating
the magefiles dependencies should ensure that the same issue is not
appearing when the operator is released.